+++ src/Psl/Hash/Hmac/hash.php
@@ -4,21 +4,18 @@ declare(strict_types=1);
 
 namespace Psl\Hash\Hmac;
 
-use Psl;
 use Psl\Hash;
 
 /**
  * Generate a keyed hash value using the HMAC method.
  *
- * @param non-empty-string $algorithm
  * @param non-empty-string $key
  *
- * @throws Psl\Exception\InvariantViolationException If the given algorithm is unsupported.
  * @throws Hash\Exception\RuntimeException If unable to pump data into the hashing context.
  *
  * @pure
  */
-function hash(string $data, string $algorithm, string $key): string
+function hash(string $data, Algorithm $algorithm, string $key): string
 {
     return Hash\Context::hmac($algorithm, $key)->update($data)->finalize();
 }