+++ src/Psl/Hash/hash.php
@@ -4,19 +4,14 @@ declare(strict_types=1);
 
 namespace Psl\Hash;
 
-use Psl;
-
 /**
  * Generate a hash value (message digest).
  *
- * @param non-empty-string $algorithm
- *
- * @throws Psl\Exception\InvariantViolationException If the given algorithm is unsupported.
  * @throws Exception\RuntimeException If unable to pump data into the hashing context.
  *
  * @pure
  */
-function hash(string $data, string $algorithm): string
+function hash(string $data, Algorithm $algorithm): string
 {
     return Context::forAlgorithm($algorithm)->update($data)->finalize();
 }